+++ QuestionsAndAnswers/server/.eslintrc.js
@@ -1,0 +1,23 @@
+module.exports = {
+  env: {
+    commonjs: true,
+    es2021: true,
+    node: true,
+  },
+  extends: [
+    'airbnb-base',
+  ],
+  parserOptions: {
+    ecmaVersion: 12,
+  },
+  rules: {
+    no-console: 'off',
+    no-restricted-syntax: [
+      "error",
+      {
+        selector: 'CallExpression[callee.object.name='console'][callee.property.name!=/^(log|warn|error|info|trace)$/]',
+        message: 'Unexpected property on console object was called'
+      }
+    ]
+  },
+};
