+++ QuestionsAndAnswers/server/index.js
@@ -1,14 +1,10 @@
 require('newrelic');
 const express = require('express');
-const routes = require('./router.js');
-const path = require('path');
-const cors = require("cors");
+const routes = require('./routes/router.js');
 
 const app = express();
 const PORT = 8080;
 
-// app.use(express.static('server'))
-app.use(cors());
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use('/', routes);
