+++ QuestionsAndAnswers/server/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "qa-api",
+  "version": "1.0.0",
+  "description": "Project-Catwalk Q&A Service - API",
+  "main": "index.js",
+  "directories": {
+    "test": "test"
+  },
+  "scripts": {
+    "start": "nodemon ./server/index.js",
+    "start:dev": "nodemon ./server/index.js --exec\"npm run lint && node\"",
+    "test:dev": "node ./test/artillery-performance-test.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/sereigh/Project-SDC.git"
+  },
+  "author": "sereigh",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/sereigh/Project-SDC/issues"
+  },
+  "homepage": "https://github.com/sereigh/Project-SDC#readme",
+  "dependencies": {
+    "express": "^4.17.1",
+    "newrelic": "^7.2.0",
+    "nodemon": "^2.0.7"
+  },
+  "devDependencies": {
+    "artillery": "^1.6.2",
+    "eslint": "^7.22.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.22.1"
+  }
+}
